# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Room Reservation",
        "description": "API Documentation for Hotel Room Reservation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/payments": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "Tampilkan semua data pembayaran (Hanya Admin)",
                "description": "Mengambil daftar lengkap semua transaksi pembayaran. Endpoint ini sebaiknya hanya dapat diakses oleh admin.",
                "operationId": "c7732b5197593b9216465b4c89957bdb",
                "responses": {
                    "200": {
                        "description": "Daftar pembayaran berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Buat pembayaran dan dapatkan Snap Token Midtrans",
                "description": "Membuat record pembayaran di database. Jika metode pembayaran 'cash', proses selesai. Jika 'online' atau kartu, akan menghasilkan snap_token dari Midtrans untuk ditampilkan di frontend.",
                "operationId": "4d18b974f35e5bcc5443060b26a3e053",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reservation_id",
                                    "amount",
                                    "payment_method"
                                ],
                                "properties": {
                                    "reservation_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 550000
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "credit_card",
                                            "debit_card",
                                            "cash",
                                            "online"
                                        ],
                                        "example": "online"
                                    },
                                    "customer_name": {
                                        "description": "Opsional, nama pelanggan.",
                                        "type": "string",
                                        "example": "John Doe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pembayaran berhasil dibuat. Jika metode online, response akan berisi 'snap_token'.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "payment": {
                                            "type": "object"
                                        },
                                        "snap_token": {
                                            "description": "Hanya muncul jika metode pembayaran bukan 'cash'.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "500": {
                        "description": "Server Error (Gagal membuat pembayaran atau token)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/payments/{id}": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "Tampilkan detail pembayaran",
                "description": "Mengambil detail satu data pembayaran berdasarkan ID.",
                "operationId": "09b5ced23a60fac5708da0d1d5f98ae2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Pembayaran",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail pembayaran"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Payment"
                ],
                "summary": "Update data pembayaran (Hanya Admin)",
                "description": "Memperbarui status atau detail pembayaran. Jika status diubah menjadi 'Paid', status reservasi terkait akan menjadi 'confirmed'. Endpoint ini hanya untuk admin.",
                "operationId": "c422013925aafb7d1e2bed4a12edfa84",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Pembayaran",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Unpaid",
                                            "Paid"
                                        ],
                                        "example": "Paid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pembayaran berhasil diupdate"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Payment"
                ],
                "summary": "Hapus data pembayaran (Hanya Admin)",
                "description": "Menghapus data pembayaran dari database. Endpoint ini hanya untuk admin.",
                "operationId": "18cbb5ddf4ff49fb0a0d914555029a43",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Pembayaran",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pembayaran berhasil dihapus"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/midtrans/callback": {
            "post": {
                "tags": [
                    "Midtrans Webhook"
                ],
                "summary": "Callback Notifikasi Pembayaran Midtrans",
                "description": "ENDPOINT INI BUKAN UNTUK DIGUNAKAN OLEH FRONTEND. Endpoint ini menerima notifikasi (webhook) dari server Midtrans untuk update status pembayaran secara otomatis.",
                "operationId": "70804a1de6576a2b99396d786d75a2ae",
                "requestBody": {
                    "description": "Payload yang dikirim oleh server Midtrans.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "transaction_time": {
                                        "type": "string",
                                        "example": "2025-06-18 09:10:00"
                                    },
                                    "transaction_status": {
                                        "type": "string",
                                        "example": "settlement"
                                    },
                                    "transaction_id": {
                                        "type": "string",
                                        "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                                    },
                                    "status_message": {
                                        "type": "string",
                                        "example": "midtrans payment notification"
                                    },
                                    "status_code": {
                                        "type": "string",
                                        "example": "200"
                                    },
                                    "signature_key": {
                                        "type": "string"
                                    },
                                    "payment_type": {
                                        "type": "string",
                                        "example": "gopay"
                                    },
                                    "order_id": {
                                        "type": "string",
                                        "example": "PAY-1-1749875460"
                                    },
                                    "merchant_id": {
                                        "type": "string"
                                    },
                                    "gross_amount": {
                                        "type": "string",
                                        "example": "550000.00"
                                    },
                                    "fraud_status": {
                                        "type": "string",
                                        "example": "accept"
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "IDR"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notifikasi berhasil diproses"
                    },
                    "404": {
                        "description": "Order ID tidak ditemukan"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}